pyproject.toml
[tool.poetry]
name = "python-django-diploma"
version = "0.1.0"
description = "This is my python django diploma"
authors = ["konstantin7297 <khromykh.kas@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.0"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
